Add package CLI entry point
